Refactor naming scheme for improved overview in tree-view and grepping
Due to Go's bs opinionated takes it's sadly not possible to prefix
either the package main files with an underscore (e.g. `route/_route.go`)
or all additional files (e.g. `route/_env.go`) to always have them sorted
at the very top of a directory when viewed in a tree-view. Hence I opted
to suffix every main file with a `._`, which however required every
additional file to be prefixed with the package name (e.g.
`route.env.go`) so that sorting works.

These changes however lead to a vastly improved overview and grepping
over the default approach that makes it hard to visually spot files in a
package.